requirements.txt
flask
pandas
numpy
redis
google-cloud-bigquery
scikit-learn
catboost
tensorflow
shap
requests
pydantic
pytest
requests-mock
pyyaml
pytest-bdd
pytest-cov
ruff
pre-commit
